+++ ontospy/ontodocs/media/templates/d3/d3_bar_hierarchy.html
@@ -1,12 +1,9 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-
-<title>{{ontology.uri}}</title>
-
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
+{% extends "d3base.html" %}
 
 
+
+
+{% block custom_css %}
 
 
 <style type="text/css" media="screen">
@@ -54,22 +51,15 @@
 
 </style>
 
+{% endblock custom_css %}
 
-</head>
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
+{% block main_content %}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{main_uri}}</a></h1>
-	<h3>{{ontology.bestDescription}}</h3>
 
-	<p><small>Ontology documentation automatically generated by <a class="greylink" href="http://lambdamusic.github.io/Ontospy/">Ontospy</a> on {% now "jS F Y H:i" %}</small></p>
-
-	<hr>
 
 
 	{% if TOTAL_CLASSES %}
@@ -93,11 +83,18 @@
 
 
 
-	<hr>
-	<small>Documentation automatically generated by <a href="https://github.com/lambdamusic/Ontospy">Ontospy</a> on {% now "jS F Y H:i" %}</small>
+{% endblock main_content %}
 
 
 
+
+
+
+
+
+
+
+{% block the_javascript %}
 
 
 
@@ -108,6 +105,17 @@
 
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery-1.7.1.min.js"></script>
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery.slidereveal.min.js"></script>
+
+
+
+
+
+
+{% include 'js_slidereveal.html' %}
+
+
+
+
 
 <script type="text/javascript">
 
@@ -343,5 +351,10 @@
 
 
 
-</body>
-</html>
+
+
+{% endblock the_javascript %}
+
+
+
+
