+++ ontospy/ontodocs/media/templates/d3/d3_bubble_chart.html
@@ -1,80 +1,67 @@
-
-
-
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-
-<title>{{ontology.uri}}</title>
-
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
+{% extends "d3base.html" %}
 
 
 
 
-<style type="text/css" media="screen">
+
+{% block custom_css %}
+
+
+  <style type="text/css" media="screen">
 
 
 
-text {
-  font-size: 11px;
-  pointer-events: none;
-}
+  text {
+    font-size: 11px;
+    pointer-events: none;
+  }
 
-text.parent {
-  fill: #1f77b4;
-}
+  text.parent {
+    fill: #1f77b4;
+  }
 
-circle {
-  fill: #ccc;
-  stroke: #999;
-  pointer-events: all;
-}
+  circle {
+    fill: #ccc;
+    stroke: #999;
+    pointer-events: all;
+  }
 
-circle.parent {
-  fill: #1f77b4;
-  fill-opacity: .1;
-  stroke: steelblue;
-}
+  circle.parent {
+    fill: #1f77b4;
+    fill-opacity: .1;
+    stroke: steelblue;
+  }
 
-circle.parent:hover {
-  stroke: #ff7f0e;
-  stroke-width: .5px;
-}
+  circle.parent:hover {
+    stroke: #ff7f0e;
+    stroke-width: .5px;
+  }
 
-{#note: June 20, 2016 removed to allow mousetips on all circles #}
-{#circle.child {#}
-{#  pointer-events: none;#}
-{# } #}
+  {#note: June 20, 2016 removed to allow mousetips on all circles #}
+  {#circle.child {#}
+  {#  pointer-events: none;#}
+  {# } #}
 
 
-circle.child:hover {
-  stroke: #ff7f0e;
-  stroke-width: .5px;
-  fill: #eaeaea;
-  fill-opacity: .9;
-}
-
+  circle.child:hover {
+    stroke: #ff7f0e;
+    stroke-width: .5px;
+    fill: #eaeaea;
+    fill-opacity: .9;
+  }
 
 
 </style>
 
+{% endblock custom_css %}
 
-</head>
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{main_uri}}</a></h1>
-	<h3>{{ontology.bestDescription}}</h3>
+{% block main_content %}
 
-	<p><small>Ontology documentation automatically generated by <a class="greylink" href="http://lambdamusic.github.io/Ontospy/">Ontospy</a> on {% now "jS F Y H:i" %}</small></p>
-
-	<hr>
 
 
 	{% if TOTAL_CLASSES %}
@@ -91,12 +78,23 @@ circle.child:hover {
 
 
 
-	<hr>
-	<small>Documentation automatically generated by <a href="https://github.com/lambdamusic/Ontospy">Ontospy</a> on {% now "jS F Y H:i" %}</small>
+  {% endblock main_content %}
 
 
 
 
+
+
+  
+
+
+
+
+
+
+
+
+{% block the_javascript %}
 
 
 
@@ -106,6 +104,16 @@ circle.child:hover {
 
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery-1.7.1.min.js"></script>
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery.slidereveal.min.js"></script>
+
+
+
+
+
+{% include 'js_slidereveal.html' %}
+
+
+
+
 
 <script type="text/javascript">
 
@@ -140,12 +148,10 @@ circle.child:hover {
 
 
      function doclick(objid){
-{#         var win = window.open("http://www.nature.com/subjects/"+objid, '_blank');#}
-{#         win.focus();#}
          console.log(objid);
+         showpanel(objid);
          };
 
-{#    d3.json("flare.json", function(error, root) {#}
     var node = svg.selectAll(".node")
           .data(bubble.nodes(classes(root))
           .filter(function(d) { return !d.children; }))
@@ -199,5 +205,6 @@ circle.child:hover {
 
 
 
-</body>
-</html>
+{% endblock the_javascript %}
+
+
