+++ ontospy/ontodocs/media/templates/d3/d3_dendogram.html
@@ -1,263 +1,183 @@
+{% extends "d3base.html" %}
 
 
 
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
 
-<title>{{ontology.uri}}</title>
-
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
-
-
+{% block custom_css %}
 
 
 <style type="text/css" media="screen">
 
 
+body {
+	margin-left: 6em;
+	margin-right: 12em;
+	font-size: 14px;
+	font-family: "Helvetica Neue", Helvetica;
+}
 
-	body {
-	    margin-left: 6em;
-	    margin-right: 12em;
-	    font-size: 14px;
-	    font-family: "Helvetica Neue", Helvetica;
-	}
+h1, h2, h3, h4, h5, h6 { text-align: left }
 
-	h1, h2, h3, h4, h5, h6 { text-align: left }
+h1, h2, h3, h4, h5 { color: #000000;}
+h2 { font: 140% sans-serif }
+h3 { font: 120% sans-serif }
+h4 { font: bold 100% sans-serif }
+h5 { font: italic 100% sans-serif }
+h6 { font: small-caps 100% sans-serif }
 
-	h1, h2, h3, h4, h5 { color: #000000;}
-	h2 { font: 140% sans-serif }
-	h3 { font: 120% sans-serif }
-	h4 { font: bold 100% sans-serif }
-	h5 { font: italic 100% sans-serif }
-	h6 { font: small-caps 100% sans-serif }
-
-	.margintop {
-		 margin-top: 40px;
-	}
-
+.margintop {
+	margin-top: 40px;
+}
 
 
-	/* rect {
-	   fill: none;
-	   pointer-events: all;
-	 }*/
 
-	 rect {
-	   fill:black;
-	   stroke:none;
-	 }
+/* rect {
+fill: none;
+pointer-events: all;
+}*/
 
-	 line {
-	   stroke: #000;
-	   stroke-width: 1.5px;
-	 }
+rect {
+fill:black;
+stroke:none;
+}
 
-	 .string, .regexp {
-	   color: #f39;
-	 }
+line {
+stroke: #000;
+stroke-width: 1.5px;
+}
 
-
-	 a:link, a:visited {
-	   color: #666;
-	   text-decoration: none;
-	 }
-
-	 a:hover {
-	   color: #000
-	 }
+.string, .regexp {
+color: #f39;
+}
 
 
-	/*extra css*/
+a:link, a:visited {
+color: #666;
+text-decoration: none;
+}
 
-	.node circle {
-	  cursor: pointer;
-	  fill: #fff;
-	  stroke: steelblue;
-	  stroke-width: 1.5px;
-	}
-
-	.node text {
-	  font-size: 11px;
-	}
-
-	path.link {
-	  fill: none;
-	  stroke: #ccc;
-	  stroke-width: 1.5px;
-	}
+a:hover {
+color: #000
+}
 
 
-	.infoentity {
-	    stroke: white;
-	    stroke-width: 2.5px;
-	    opacity: 0.7;
-		background: linen;
-		line-height: 0.7em;
-	}
+/*extra css*/
+
+.node circle {
+cursor: pointer;
+fill: #fff;
+stroke: steelblue;
+stroke-width: 1.5px;
+}
+
+.node text {
+font-size: 11px;
+}
+
+path.link {
+fill: none;
+stroke: #ccc;
+stroke-width: 1.5px;
+}
 
 
-	.smalldesc {
-		font-size: 20px;
-		margin-top: 0px;
-		color: black;
-	}
+.infoentity {
+	stroke: white;
+	stroke-width: 2.5px;
+	opacity: 0.7;
+	background: linen;
+	line-height: 0.7em;
+}
 
 
-	.hidden {
-		display: none;
-	}
+.smalldesc {
+	font-size: 20px;
+	margin-top: 0px;
+	color: black;
+}
 
-	kbd {
-		padding: 2px 4px;
-	    font-size: 90%;
-	    color: #fff;
-	    background-color: #333;
-	    border-radius: 3px;
-	    box-shadow: inset 0 -1px 0 rgba(0,0,0,.25);border-image-outset:
-	}
+kbd {
+	padding: 2px 4px;
+	font-size: 90%;
+	color: #fff;
+	background-color: #333;
+	border-radius: 3px;
+	box-shadow: inset 0 -1px 0 rgba(0,0,0,.25);border-image-outset:
+}
+    
 
-	/*slider*/
-
-	.side-panel{
-		background-color: #E6E6E6;
-		padding: 10px;
-		/*word-wrap: break-word;*/
-	    overflow: auto;
-	}
-
-	.side-panel-left{
-		box-shadow: -5px 0 15px -6px #585858 inset;
-	}
-
-	.side-panel-right{
-		box-shadow: 5px 0 15px -6px #585858 inset;
-	}
-
-	a.class {
-		color: darkred;
-	}
-	a.prop {
-		color: darkblue;
-	}
-
-	.turtlesource {
-		font-size: 12px;
-	}
-
-	dt {
-		color: darkslategrey;
-	}
-
-	.text-muted {
-		color: grey;
-	}
-
-	hr {
-		margin-top: 20px;
-	    margin-bottom: 20px;
-	}
-
-    .credits {
-        color: gray;
-    }
-
-    a.greylink {
-        color: #666;
-        text-decoration: underline;
-    }
-
-    .section-title {
-{#        text-transform: uppercase;#}
-        font-size: smaller;
-    }
 
 </style>
 
 
-</head>
+{% endblock custom_css %}
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{docs_title}}</a></h1>
 
-	<p>
-    <small class="credits">Ontology documentation automatically generated with <a class="greylink" href="http://lambdamusic.github.io/Ontospy/" target="_blank">Ontospy</a> on {% now "jS F Y H:i" %} </small>
-  </p>
 
-	<hr>
+{% block main_content %}
+
 
 
 	{% if TOTAL_CLASSES %}
-		<div class="row">
-			<h3>Classes ({{TOTAL_CLASSES}})</h3>
-			<div class="col-md-12" id="graph_classes">
-
-
-			</div>
+	<div class="row">
+		<h3>Classes ({{TOTAL_CLASSES}})</h3>
+		<div class="col-md-12" id="graph_classes">
 
 
 		</div>
+
+
+	</div>
 	{% endif %}
 
 	{% if TOTAL_PROPERTIES %}
-		<div class="row">
-			<h3>Properties ({{TOTAL_PROPERTIES}})</h3>
-			<div class="col-md-12" id="graph_properties">
-
-
-			</div>
+	<div class="row">
+		<h3>Properties ({{TOTAL_PROPERTIES}})</h3>
+		<div class="col-md-12" id="graph_properties">
 
 
 		</div>
+
+
+	</div>
 	{% endif %}
 
 
 	{% if TOTAL_CONCEPTS %}
 
-		<div class="row">
-			<h3>Concepts ({{TOTAL_CONCEPTS}})</h3>
-			<div class="col-md-12" id="graph_concepts">
-
-
-			</div>
+	<div class="row">
+		<h3>Concepts ({{TOTAL_CONCEPTS}})</h3>
+		<div class="col-md-12" id="graph_concepts">
 
 
 		</div>
+
+
+	</div>
 
 	{% endif %}
 
 
 
 
-	<hr>
-    <small class="credits">Ontology documentation automatically generated with <a class="greylink" href="http://lambdamusic.github.io/Ontospy/" target="_blank">Ontospy</a> on {% now "jS F Y H:i" %} </small>
-
-
-
-
-	{# SLIDER AND HIDDEN TABS #}
-
-	{% include "d3tree_hiddendivs.html" %}
-
-
-	<div id='slider' class="side-panel side-panel-left">
-		<span class="text-muted">Press <kbd>ESC</kbd> to close</span>
-		<div id="slide_contents"></div>
-	</div>
-
-	{# <button id='trigger'>Trigger</button> #}
+{% endblock main_content %}
 
 
 
 
 
-{# <script type="text/javascript" src="static/d3.layout.tree.js"></script> #}
+
+
+
+
+
+{% block the_javascript %}
+
+
 
 
 
@@ -286,31 +206,7 @@ https://cdnjs.cloudflare.com/ajax/libs/jquery/1.7.1/jquery.min.js"></script>
 
 
 
-    <script type="text/javascript">
-
-		var slider = $("#slider").slideReveal({
-			  trigger: $("#trigger"),
-			  push: false,
-			  overlay: true,
-			  width: '50%',
-			  speed: 700 ,
-			});
-
-		function showpanel(argument) {
-			console.log("#"+argument);
-			if (argument != "None") {
-				$("#slide_contents").empty();
-				$("#slide_contents").html($("#"+argument).html());
-				slider.slideReveal("show");
-			}
-		}
-
-
-
-
-
-    </script>
-
+	{% include 'js_slidereveal.html' %}
 
 
 
@@ -353,7 +249,6 @@ https://cdnjs.cloudflare.com/ajax/libs/jquery/1.7.1/jquery.min.js"></script>
 	  	.append("svg:g")
 	    .attr("transform", "translate(" + m[3] + "," + m[0] + ")");
 
-	// d3.json(JSON_FILE, function(json) {
 	root_classes = json1;
 	root_classes.x0 = h / 2;
 	root_classes.y0 = 0;
@@ -363,12 +258,14 @@ https://cdnjs.cloudflare.com/ajax/libs/jquery/1.7.1/jquery.min.js"></script>
 
 		// Initialize the display to show no nodes:
 		root_classes.children.forEach(toggleAll);
+
 	{% endif %}
 
 
 	update(vis1, tree1, diagonal1, root_classes, json1);
 
 {% endif %}
+
 
 
 
@@ -408,7 +305,6 @@ https://cdnjs.cloudflare.com/ajax/libs/jquery/1.7.1/jquery.min.js"></script>
 	  	.append("svg:g")
 	    .attr("transform", "translate(" + m[3] + "," + m[0] + ")");
 
-	// d3.json(JSON_FILE, function(json) {
 	root_properties = json2;
 	root_properties.x0 = h / 2;
 	root_properties.y0 = 0;
@@ -651,8 +547,8 @@ function toggle(d) {
 
 
 
+{% endblock the_javascript %}
 
 
 
-</body>
-</html>
+
