+++ ontospy/ontodocs/media/templates/d3/d3_pack_hierarchy.html
@@ -1,15 +1,9 @@
+{% extends "d3base.html" %}
 
 
 
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
 
-<title>{{ontology.uri}}</title>
-
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
-
-
+{% block custom_css %}
 
 
 <style type="text/css" media="screen">
@@ -57,114 +51,56 @@ circle.child:hover {
 
 
 
-	/*extra css*/
+/*extra css*/
 
-	.node circle {
-	  cursor: pointer;
-	  fill: #fff;
-	  stroke: steelblue;
-	  stroke-width: 1.5px;
-	}
+.node circle {
+	cursor: pointer;
+	fill: #fff;
+	stroke: steelblue;
+	stroke-width: 1.5px;
+}
 
-	.node text {
-	  font-size: 11px;
-	}
+.node text {
+	font-size: 11px;
+}
 
-	path.link {
-	  fill: none;
-	  stroke: #ccc;
-	  stroke-width: 1.5px;
-	}
-
-
-	.infoentity {
-	    stroke: white;
-	    stroke-width: 2.5px;
-	    opacity: 0.7;
-		background: linen;
-		line-height: 0.7em;
-	}
+path.link {
+	fill: none;
+	stroke: #ccc;
+	stroke-width: 1.5px;
+}
 
 
-	.smalldesc {
-		font-size: 20px;
-		margin-top: 0px;
-		color: black;
-	}
+.infoentity {
+	stroke: white;
+	stroke-width: 2.5px;
+	opacity: 0.7;
+	background: linen;
+	line-height: 0.7em;
+}
 
 
-	.hidden {
-		display: none;
-	}
-
-	kbd {
-		padding: 2px 4px;
-	    font-size: 90%;
-	    color: #fff;
-	    background-color: #333;
-	    border-radius: 3px;
-	    box-shadow: inset 0 -1px 0 rgba(0,0,0,.25);border-image-outset:
-	}
-
-	/*slider*/
-
-	.side-panel{
-		background-color: #E6E6E6;
-		padding: 10px;
-		/*word-wrap: break-word;*/
-	    overflow: auto;
-	}
-
-	.side-panel-left{
-		box-shadow: -5px 0 15px -6px #585858 inset;
-	}
-
-	.side-panel-right{
-		box-shadow: 5px 0 15px -6px #585858 inset;
-	}
-
-	a.class {
-		color: darkred;
-	}
-	a.prop {
-		color: darkblue;
-	}
-
-	.turtlesource {
-		font-size: 12px;
-	}
-
-	dt {
-		color: darkslategrey;
-	}
-
-	.text-muted {
-		color: grey;
-	}
-
-	hr {
-		margin-top: 20px;
-	    margin-bottom: 20px;
-	}
+.smalldesc {
+	font-size: 20px;
+	margin-top: 0px;
+	color: black;
+}
 
 </style>
 
 
-</head>
+{% endblock custom_css %}
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{main_uri}}</a></h1>
-	<h3>{{ontology.bestDescription}}</h3>
 
-	<p><small>Ontology documentation automatically generated by <a class="greylink" href="http://lambdamusic.github.io/Ontospy/">Ontospy</a> on {% now "jS F Y H:i" %}</small></p>
 
-	<hr>
+
+{% block main_content %}
+
+
 
 
 	{% if TOTAL_CLASSES %}
@@ -181,10 +117,13 @@ circle.child:hover {
 
 
 
-	<hr>
-	<small>Documentation automatically generated by <a href="https://github.com/lambdamusic/Ontospy">Ontospy</a> on {% now "jS F Y H:i" %}</small>
+{% endblock main_content %}
 
 
+
+
+
+{% block the_javascript %}
 
 
 
@@ -194,6 +133,12 @@ circle.child:hover {
 
   <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery-1.7.1.min.js"></script>
   <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery.slidereveal.min.js"></script>
+
+
+
+	{% include 'js_slidereveal.html' %}
+
+
 
 
 <script type="text/javascript">
@@ -310,5 +255,7 @@ circle.child:hover {
 </script>
 
 
-</body>
-</html>
+
+{% endblock the_javascript %}
+
+
