+++ ontospy/ontodocs/media/templates/d3/d3_rotating_cluster.html
@@ -1,12 +1,9 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-
-<title>{{ontology.uri}}</title>
-
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
+{% extends "d3base.html" %}
 
 
+
+
+{% block custom_css %}
 
 
 <style type="text/css" media="screen">
@@ -79,21 +76,15 @@ circle.child:hover {
 </style>
 
 
-</head>
+{% endblock custom_css %}
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{main_uri}}</a></h1>
-	<h3>{{ontology.bestDescription}}</h3>
 
-	<p><small>Ontology documentation automatically generated by <a class="greylink" href="http://lambdamusic.github.io/Ontospy/">Ontospy</a> on {% now "jS F Y H:i" %}</small></p>
 
-	<hr>
+  {% block main_content %}
 
 
 	{% if TOTAL_CLASSES %}
@@ -108,13 +99,13 @@ circle.child:hover {
 		</div>
 	{% endif %}
 
-
-
-	<hr>
-	<small>Documentation automatically generated by <a href="https://github.com/lambdamusic/Ontospy">Ontospy</a> on {% now "jS F Y H:i" %}</small>
+  {% endblock main_content %}
 
 
 
+
+
+  {% block the_javascript %}
 
 
 
@@ -125,6 +116,17 @@ circle.child:hover {
 
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery-1.7.1.min.js"></script>
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery.slidereveal.min.js"></script>
+
+
+
+
+
+{% include 'js_slidereveal.html' %}
+
+
+
+
+
 
 <script type="text/javascript">
 
@@ -248,5 +250,10 @@ circle.child:hover {
 
 
 
-</body>
-</html>
+
+
+
+    {% endblock the_javascript %}
+
+
+
