+++ ontospy/ontodocs/media/templates/d3/d3_treePie.html
@@ -1,11 +1,9 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
+{% extends "d3base.html" %}
 
-<title>{{ontology.uri}}</title>
 
-<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
 
+
+{% block custom_css %}
 
 
 
@@ -19,11 +17,6 @@
     body
     {
         padding:10px;
-    }
-
-    form
-    {
-
     }
 
     #sb_hierarchy
@@ -44,21 +37,21 @@
 </style>
 
 
-</head>
+{% endblock custom_css %}
 
 
-<body>
 
-	{# VISIBLE PAGE HTML #}
 
-	{# note: showpanel arg is the ID of the div that is showing.. #}
 
-	<h1 class="title">Ontology:&nbsp;<a href="javascript:showpanel('ontoinfo');">{{main_uri}}</a></h1>
-	<h3>{{ontology.bestDescription}}</h3>
 
-	<p><small>Ontology documentation automatically generated by <a class="greylink" href="http://lambdamusic.github.io/Ontospy/">Ontospy</a> on {% now "jS F Y H:i" %}</small></p>
 
-	<hr>
+
+
+
+{% block main_content %}
+
+
+
 
 
 	{% if TOTAL_CLASSES %}
@@ -76,12 +69,17 @@
 	{% endif %}
 
 
-
-	<hr>
-	<small>Documentation automatically generated by <a href="https://github.com/lambdamusic/Ontospy">Ontospy</a> on {% now "jS F Y H:i" %}</small>
+{% endblock main_content %}
 
 
 
+
+
+
+
+
+
+{% block the_javascript %}
 
 
 
@@ -92,6 +90,15 @@
 
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery-1.7.1.min.js"></script>
 <script type="text/javascript" src="{{STATIC_URL}}libs/jquery/jquery.slidereveal.min.js"></script>
+
+
+
+{% include 'js_slidereveal.html' %}
+
+
+
+
+
 
 <script type="text/javascript">
 
@@ -367,5 +374,10 @@
 
 
 
-</body>
-</html>
+
+
+	{% endblock the_javascript %}
+
+
+
+
