+++ ontospy/ontodocs/viz/viz_d3rotating_cluster.py
@@ -10,6 +10,8 @@ from ..viz_factory import VizFactory
 
 # ===========
 # D3 ROTATING CLUSTER
+# 
+# NOTE results are not that good 
 # ===========
 
 
@@ -48,7 +50,7 @@ class Dataviz(VizFactory):
 
         # Ontology - MAIN PAGE
         contents = self._renderTemplate(
-            "d3/d3_rotatingCluster.html", extraContext=extra_context)
+            "d3/d3_rotating_cluster.html", extraContext=extra_context)
         FILE_NAME = "index.html"
         main_url = self._save2File(contents, FILE_NAME, self.output_path)
 
@@ -73,54 +75,3 @@ if __name__ == '__main__':
     except KeyboardInterrupt as e:  # Ctrl-C
         raise e
 
-
-def run(graph, save_on_github=False, main_entity=None):
-    """
-	"""
-    try:
-        ontology = graph.all_ontologies[0]
-        uri = ontology.uri
-    except:
-        ontology = None
-        uri = ";".join([s for s in graph.sources])
-
-    # ontotemplate = open("template.html", "r")
-    ontotemplate = open(ONTODOCS_VIZ_TEMPLATES + "d3_cluster.html", "r")
-    t = Template(ontotemplate.read())
-
-    c_total = len(graph.classes)
-
-    mylist = build_D3treeStandard(0, 99, 1, graph.toplayer_classes)
-
-    JSON_DATA_CLASSES = json.dumps({
-        'children': mylist,
-        'name': 'owl:Thing',
-    })
-
-    c = Context({
-        "ontology": ontology,
-        "main_uri": uri,
-        "STATIC_PATH": ONTODOCS_VIZ_STATIC,
-        "save_on_github": save_on_github,
-        'JSON_DATA_CLASSES': JSON_DATA_CLASSES,
-        "TOTAL_CLASSES": c_total,
-    })
-
-    rnd = t.render(c)
-
-    return safe_str(rnd)
-
-
-if __name__ == '__main__':
-    import sys
-    try:
-        # script for testing - must launch this module
-        # >python -m viz.viz_packh
-
-        func = locals()["run"]  # main func dynamically
-        run_test_viz(func)
-
-        sys.exit(0)
-
-    except KeyboardInterrupt as e:  # Ctrl-C
-        raise e
