+++ ontospy/ontodocs/viz/viz_d3tree_pie.py
@@ -1,6 +1,19 @@
 # !/usr/bin/env python
 #  -*- coding: UTF-8 -*-
 
+
+# ===========
+# D3 TREE PIE
+#
+# NOT IN USE  / NEEDS MORE TESTING
+# ===========
+
+
+
+
+
+
+
 import os, sys
 import json
 
@@ -8,37 +21,34 @@ from ..utils import *
 from ..builder import *  # loads and sets up Django
 from ..viz_factory import VizFactory
 
-# ===========
-# D3 BUBBLE CHART
-# ===========
 
 
 class Dataviz(VizFactory):
     """
-    D3 Bubbles
+	D3 TREE PIE
 
-    """
+	"""
 
     def __init__(self, ontospy_graph, title=""):
         """
-        Init
-        """
+		Init
+		"""
         super(Dataviz, self).__init__(ontospy_graph, title)
         self.static_files = ["libs/d3-v3", "libs/jquery"]
 
     def _buildTemplates(self):
         """
-        OVERRIDING THIS METHOD from Factory
-        """
+		OVERRIDING THIS METHOD from Factory
+		"""
 
-        jsontree_classes = build_D3bubbleChart(
-            0, 99, 1, self.ontospy_graph.toplayer_classes)
+        jsontree_classes = build_D3treepie(0, 99, 1,
+                                           self.ontospy_graph.toplayer_classes)
         c_total = len(self.ontospy_graph.all_classes)
+        c_toplayer = len(self.ontospy_graph.toplayer_classes)
 
-        JSON_DATA_CLASSES = json.dumps({
-            'children': jsontree_classes,
-            'name': 'owl:Thing',
-        })
+        # weird - DBCheck!
+        JSON_DATA_CLASSES = json.dumps(
+            ["owl:Thing", [c_toplayer, c_toplayer], jsontree_classes])
 
         extra_context = {
             "ontograph": self.ontospy_graph,
@@ -48,7 +58,7 @@ class Dataviz(VizFactory):
 
         # Ontology - MAIN PAGE
         contents = self._renderTemplate(
-            "d3/d3_bubblechart.html", extraContext=extra_context)
+            "d3/d3_treepie.html", extraContext=extra_context)
         FILE_NAME = "index.html"
         main_url = self._save2File(contents, FILE_NAME, self.output_path)
